# Fix memory issue & improve processing speed