--- a/cache/cache.go
+++ b/cache/cache.go
@@ -24,7 +24,7 @@ type ContentMetadata struct {
 
 type CachedData struct {
 	ContentMetadata
-	Data io.ReadCloser
+	Data io.ReadCloser // we need a ReadCloser because the reader is used oustide the scope where it was created and need to be closed by the function using it.
 	Ttl  time.Duration
 }