--- a/cache/filesystem_cache.go
+++ b/cache/filesystem_cache.go
@@ -1,7 +1,6 @@
 package cache
 
 import (
-	"bytes"
 	"fmt"
 	"io"
 	"math/rand"
@@ -101,7 +100,8 @@ func (f *fileSystemCache) Get(key *Key) (*CachedData, error) {
 		return nil, ErrMissing
 	}
 
-	defer file.Close()
+	// the file will be closed once it's read as an io.ReaderCloser
+	// This  ReaderCloser is stored in the returned CachedData
 	fi, err := file.Stat()
 	if err != nil {
 		return nil, fmt.Errorf("cache %q: cannot stat %q: %w", f.Name(), fp, err)
@@ -111,28 +111,26 @@ func (f *fileSystemCache) Get(key *Key) (*CachedData, error) {
 	if age > f.expire {
 		// check if file exceeded expiration time + grace time
 		if age > f.expire+f.grace {
+			file.Close()
 			return nil, ErrMissing
 		}
 		// Serve expired file in the hope it will be substituted
 		// with the fresh file during deadline.
 	}
 
-	b, err := io.ReadAll(file)
-
 	if err != nil {
+		file.Close()
 		return nil, fmt.Errorf("failed to read file content from %q: %w", f.Name(), err)
 	}
 
-	reader := bytes.NewReader(b)
-
-	metadata, err := decodeHeader(reader)
+	metadata, err := decodeHeader(file)
 	if err != nil {
 		return nil, err
 	}
 
 	value := &CachedData{
 		ContentMetadata: *metadata,
-		Data:            reader,
+		Data:            file,
 		Ttl:             f.expire - age,
 	}
 
@@ -141,7 +139,7 @@ func (f *fileSystemCache) Get(key *Key) (*CachedData, error) {
 
 // decodeHeader decodes header from raw byte stream. Data is encoded as follows:
 // length(contentType)|contentType|length(contentEncoding)|contentEncoding|length(contentLength)|contentLength|cachedData
-func decodeHeader(reader *bytes.Reader) (*ContentMetadata, error) {
+func decodeHeader(reader io.Reader) (*ContentMetadata, error) {
 	contentType, err := readHeader(reader)
 	if err != nil {
 		return nil, fmt.Errorf("cannot read Content-Type from provided reader: %w", err)