--- a/cache/filesystem_cache.go
+++ b/cache/filesystem_cache.go
@@ -16,10 +16,6 @@ import (
 	"github.com/contentsquare/chproxy/log"
 )
 
-// Version must be increased with each backward-incompatible change
-// in the cache storage.
-const Version = 3
-
 var cachefileRegexp = regexp.MustCompile(`^[0-9a-f]{32}$`)
 
 // fileSystemCache represents a file cache.