--- a/cache/redis_cache.go
+++ b/cache/redis_cache.go
@@ -3,9 +3,8 @@ package cache
 import (
 	"bytes"
 	"context"
-	"encoding/base64"
-	"encoding/json"
 	"errors"
+	"fmt"
 	"io"
 	"regexp"
 	"strconv"
@@ -25,13 +24,6 @@ type redisCache struct {
 const getTimeout = 1 * time.Second
 const putTimeout = 2 * time.Second
 const statsTimeout = 500 * time.Millisecond
-
-type redisCachePayload struct {
-	Length   int64  `json:"l"`
-	Type     string `json:"t"`
-	Encoding string `json:"enc"`
-	Payload  string `json:"payload"`
-}
 
 func newRedisCache(client redis.UniversalClient, cfg config.Cache) *redisCache {
 	redisCache := &redisCache{
@@ -95,7 +87,6 @@ func (r *redisCache) Get(key *Key) (*CachedData, error) {
 	ctx, cancelFunc := context.WithTimeout(context.Background(), getTimeout)
 	defer cancelFunc()
 	val, err := r.client.Get(ctx, key.String()).Result()
-
 	// if key not found in cache
 	if errors.Is(err, redis.Nil) {
 		return nil, ErrMissing
@@ -106,35 +97,17 @@ func (r *redisCache) Get(key *Key) (*CachedData, error) {
 		log.Errorf("failed to get key %s with error: %s", key.String(), err)
 		return nil, ErrMissing
 	}
-
-	var payload redisCachePayload
-	err = json.Unmarshal([]byte(val), &payload)
-
-	if err != nil {
-		log.Errorf("corrupted payload for key %s with error: %s", key.String(), err)
-		return nil, ErrMissing
-	}
-
 	ttl, err := r.client.TTL(ctx, key.String()).Result()
-
 	if err != nil {
-		log.Errorf("Not able to fetch TTL for: %s ", key)
+		return nil, fmt.Errorf("failed to ttl of key %s with error: %s", key.String(), err)
 	}
+	content, reader := r.fromByte([]byte(val))
 
-	decoded, err := base64.StdEncoding.DecodeString(payload.Payload)
-	if err != nil {
-		log.Errorf("failed to decode payload: %s , due to: %v ", payload.Payload, err)
-		return nil, ErrMissing
-	}
-	reader := &io_reader_decorator{Reader: bytes.NewReader(decoded)}
+	reader2 := &io_reader_decorator{Reader: reader}
 	value := &CachedData{
-		ContentMetadata: ContentMetadata{
-			Length:   payload.Length,
-			Type:     payload.Type,
-			Encoding: payload.Encoding,
-		},
-		Data: reader,
-		Ttl:  ttl,
+		ContentMetadata: *content,
+		Data:            reader2,
+		Ttl:             ttl,
 	}
 
 	return value, nil
@@ -149,27 +122,63 @@ type io_reader_decorator struct {
 func (m io_reader_decorator) Close() error {
 	return nil
 }
+func (r *redisCache) stringToBytes(s string) []byte {
+	n := uint32(len(s))
+
+	b := make([]byte, 0, n+4)
+	b = append(b, byte(n>>24), byte(n>>16), byte(n>>8), byte(n))
+	b = append(b, s...)
+	return b
+}
+
+func (r *redisCache) stringFromBytes(bytes []byte) (string, int) {
+	b := bytes[:4]
+	n := uint32(b[3]) | (uint32(b[2]) << 8) | (uint32(b[1]) << 16) | (uint32(b[0]) << 24)
+	s := bytes[4 : 4+n]
+	return string(s), int(4 + n)
+}
+
+func (r *redisCache) toByte(contentMetadata *ContentMetadata, reader io.Reader) ([]byte, error) {
+	data, err := toBytes(reader)
+	if err != nil {
+		return nil, err
+	}
+	cLength := contentMetadata.Length
+	cType := r.stringToBytes(contentMetadata.Type)
+	cEncoding := r.stringToBytes(contentMetadata.Encoding)
+	b := make([]byte, 0, len(data)+len(cEncoding)+len(cType)+8)
+	b = append(b, byte(cLength>>56), byte(cLength>>48), byte(cLength>>40), byte(cLength>>32), byte(cLength>>24), byte(cLength>>16), byte(cLength>>8), byte(cLength))
+	b = append(b, cType...)
+	b = append(b, cEncoding...)
+	b = append(b, data...)
+	return b, nil
+}
+
+func (r *redisCache) fromByte(b []byte) (*ContentMetadata, io.Reader) {
+	cLength := uint64(b[7]) | (uint64(b[6]) << 8) | (uint64(b[5]) << 16) | (uint64(b[4]) << 24) | uint64(b[3])<<32 | (uint64(b[2]) << 40) | (uint64(b[1]) << 48) | (uint64(b[0]) << 56)
+	offset := 8
+	cType, sizeCType := r.stringFromBytes(b[offset:])
+	offset += sizeCType
+	cEncoding, sizeCEncoding := r.stringFromBytes(b[offset:])
+	offset += sizeCEncoding
+	payload := b[offset:]
+	metadata := &ContentMetadata{
+		Length:   int64(cLength),
+		Type:     cType,
+		Encoding: cEncoding,
+	}
+	return metadata, bytes.NewReader(payload)
+}
 
 func (r *redisCache) Put(reader io.Reader, contentMetadata ContentMetadata, key *Key) (time.Duration, error) {
-	data, err := toBytes(reader)
+
+	payload, err := r.toByte(&contentMetadata, reader)
 	if err != nil {
 		return 0, err
 	}
-
-	encoded := base64.StdEncoding.EncodeToString(data)
-	payload := &redisCachePayload{
-		Length: contentMetadata.Length, Type: contentMetadata.Type, Encoding: contentMetadata.Encoding, Payload: encoded,
-	}
-
-	marshalled, err := json.Marshal(payload)
-	if err != nil {
-		return 0, nil
-	}
-
 	ctx, cancelFunc := context.WithTimeout(context.Background(), putTimeout)
 	defer cancelFunc()
-	err = r.client.Set(ctx, key.String(), marshalled, r.expire).Err()
-
+	err = r.client.Set(ctx, key.String(), payload, r.expire).Err()
 	if err != nil {
 		return 0, err
 	}
@@ -182,11 +191,17 @@ func (r *redisCache) Name() string {
 }
 
 func toBytes(stream io.Reader) ([]byte, error) {
-	buf := new(bytes.Buffer)
+	b, err := io.ReadAll(stream)
+	if err != nil {
+		return nil, err
+	}
+	return b, nil
+
+	/*buf := new(bytes.Buffer)
 
 	_, err := buf.ReadFrom(stream)
 	if err != nil {
 		return nil, err
 	}
-	return buf.Bytes(), nil
+	return buf.Bytes(), nil*/
 }