## Description
There is a huge memory overhead when caching data in either redis or fscache (it's a regression introduce during the redis cache feature).
It was raised by 2 people in this issue 
https://github.com/ContentSquare/chproxy/issues/206
and this PR
https://github.com/ContentSquare/chproxy/pull/210

The aim of this PR is to fix it and improve (if possible) the query processing speed.

Here is a summary of the improvements.
All the scenarios are improved when cache is on:
- scenario 1: file system cache on cache miss
- scenario 2: file system cache on cache hit
- scenario 3: redis cache on cache miss
- scenario 4 redis cache on cache hit

And here are the result of a bench on my laptop (that consist of running sequentially 20 queries that fetch 64MB of data each from clickhouse)
before fix:
- scenario 1: heap size: 150MB | processing speed: 220MB/sec then* 70MB/sec 
- scenario 2: heap size: 150MB | processing speed: 1.2GB/sec  
- scenario 3: heap size: 750MB |  processing speed: 120MB/sec then* 80MB/sec 
- scenario 4:  heap size: 350MB |  processing speed: 140MB/sec then* 70MB/sec 

after fix
- scenario 1: heap size: 10MB** | processing speed: 220MB/sec then* 70MB/sec 
- scenario 2: heap size: 10MB** | processing speed: 3.1GB/sec 
- scenario 3: heap size: 15MB** |  processing speed: 140MB/sec
- scenario 4: heap size: 17MB** |  processing speed: 1,2GB/sec

** The size of the heap at startup is 7MB
*The throughput decreases a lot after 5-8 queries, it's likely because my ssd disk speed degrades on bursts (because of some internal caches I guess) or because of the turbo boost of intel CPUs.
## Pull request type

<!-- Please do not submit updates to dependencies unless it fixes an issue. --> 

<!-- Please try to limit your pull request to one type, submit multiple pull requests if needed. --> 

Please check the type of change your PR introduces:
- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):


### Checklist

- [ ] Linter passes correctly
- [ ] Add tests which fail without the change (if possible)
- [x] All tests passing
- [ ] Extended the README / documentation, if necessary

## Does this introduce a breaking change?

<!-- If this introduces a breaking change, please describe the impact and migration path for existing applications below. -->

- [X] Yes
- [ ] No
This PR will make the objects store in fs & redis cache in the previous version unreachable because the key generation has changed since the format is different
## Further comments

<!-- If this is a relatively large or complex change, kick off the discussion by explaining why you chose the solution you did and what alternatives you considered, etc... -->